test(auth): add dedicated device poll tests
Poll tests existed but were co-located in device_auth_test.go. Extracted
them into device_poll_test.go with fuller coverage: whitespace device
code -> ErrInvalidInput, and unexpected-error pass-through.
