+++ internal/app/usecases/auth/device_auth_test.go
@@ -67,96 +67,6 @@ func TestRequestDeviceAuthUseCase_Execute_LogtoError(t *testing.T) {
 	output, err := uc.Execute(context.Background(), auth.RequestDeviceAuthInput{
 		Email: "user@example.com",
 	})
-
 	require.Error(t, err)
-	require.Nil(t, output)
-}
-
-// ---------------------------------------------------------------------------
-// PollDeviceTokenUseCase
-// ---------------------------------------------------------------------------
-
-func TestPollDeviceTokenUseCase_Execute_Success(t *testing.T) {
-	logtoDevice := auth.NewMockLogtoDeviceFlowClient(t)
-
-	logtoDevice.EXPECT().PollDeviceToken(mock.Anything, "device_code_123").
-		Return(&logto.DeviceTokenResponse{
-			AccessToken:  "access_token_abc",
-			IDToken:      "id_token_def",
-			RefreshToken: "refresh_token_ghi",
-			ExpiresIn:    3600,
-			TokenType:    "Bearer",
-		}, nil)
-
-	uc := auth.NewPollDeviceTokenUseCase(logtoDevice)
-	output, err := uc.Execute(context.Background(), auth.PollDeviceTokenInput{
-		DeviceCode: "device_code_123",
-	})
-
-	require.NoError(t, err)
-	require.NotNil(t, output)
-	assert.Equal(t, "access_token_abc", output.AccessToken)
-	assert.Equal(t, "id_token_def", output.IDToken)
-	assert.Equal(t, "refresh_token_ghi", output.RefreshToken)
-	assert.Equal(t, 3600, output.ExpiresIn)
-}
-
-func TestPollDeviceTokenUseCase_Execute_Pending(t *testing.T) {
-	logtoDevice := auth.NewMockLogtoDeviceFlowClient(t)
-
-	logtoDevice.EXPECT().PollDeviceToken(mock.Anything, "device_code_123").
-		Return(nil, logto.ErrDeviceCodePending)
-
-	uc := auth.NewPollDeviceTokenUseCase(logtoDevice)
-	output, err := uc.Execute(context.Background(), auth.PollDeviceTokenInput{
-		DeviceCode: "device_code_123",
-	})
-
-	require.Error(t, err)
-	require.ErrorIs(t, err, logto.ErrDeviceCodePending)
-	require.Nil(t, output)
-}
-
-func TestPollDeviceTokenUseCase_Execute_Expired(t *testing.T) {
-	logtoDevice := auth.NewMockLogtoDeviceFlowClient(t)
-
-	logtoDevice.EXPECT().PollDeviceToken(mock.Anything, "device_code_expired").
-		Return(nil, logto.ErrDeviceCodeExpired)
-
-	uc := auth.NewPollDeviceTokenUseCase(logtoDevice)
-	output, err := uc.Execute(context.Background(), auth.PollDeviceTokenInput{
-		DeviceCode: "device_code_expired",
-	})
-
-	require.Error(t, err)
-	require.ErrorIs(t, err, logto.ErrDeviceCodeExpired)
-	require.Nil(t, output)
-}
-
-func TestPollDeviceTokenUseCase_Execute_AccessDenied(t *testing.T) {
-	logtoDevice := auth.NewMockLogtoDeviceFlowClient(t)
-
-	logtoDevice.EXPECT().PollDeviceToken(mock.Anything, "device_code_denied").
-		Return(nil, logto.ErrDeviceCodeAccessDenied)
-
-	uc := auth.NewPollDeviceTokenUseCase(logtoDevice)
-	output, err := uc.Execute(context.Background(), auth.PollDeviceTokenInput{
-		DeviceCode: "device_code_denied",
-	})
-
-	require.Error(t, err)
-	require.ErrorIs(t, err, logto.ErrDeviceCodeAccessDenied)
-	require.Nil(t, output)
-}
-
-func TestPollDeviceTokenUseCase_Execute_EmptyDeviceCode(t *testing.T) {
-	logtoDevice := auth.NewMockLogtoDeviceFlowClient(t)
-	uc := auth.NewPollDeviceTokenUseCase(logtoDevice)
-	output, err := uc.Execute(context.Background(), auth.PollDeviceTokenInput{
-		DeviceCode: "",
-	})
-
-	require.Error(t, err)
-	require.ErrorIs(t, err, errs.ErrInvalidInput)
 	require.Nil(t, output)
 }
